Support multiple language game data (#469)
* Support multiple language game data

- Created `xiv-gen-macros` crate with `FromCsv` derive macro for efficient game data parsing.
- Refactored `xiv-gen` to use manually defined structs with only the required fields, reducing data size.
- Updated `xiv-gen-db` to generate and optionally embed compressed bincode files for English, Japanese, German, and French.
- Updated `ultros` backend to serve language-specific game data via `/static/data/:version/:lang`.
- Updated `ultros-client` to fetch language-specific bincode files.
- Fixed various compilation errors resulting from the transition to manual struct definitions.

The application now supports multiple data versions and languages, providing a foundation for high-level language swapping.


* Support multiple language game data and fix CI

- Created `xiv-gen-macros` crate with `FromCsv` derive macro for efficient game data parsing.
- Refactored `xiv-gen` to use manually defined structs with only the required fields, reducing data size.
- Updated `xiv-gen-db` to generate and optionally embed compressed bincode files for English, Japanese, German, and French.
- Updated `ultros` backend to serve language-specific game data via `/static/data/:version/:lang`.
- Updated `ultros-client` to fetch language-specific bincode files.
- Fixed `cargo fmt` violations in `ultros/src/web.rs`.
- Updated `xiv-gen-db/build.rs` to skip CN and KO languages due to differing CSV structures.
- Fixed `xiv-gen-db` build script panic by correctly locating CSV files using `CARGO_MANIFEST_DIR`.
- Resolved compilation errors from the transition to manual struct definitions.


* Support multiple language game data

- Created `xiv-gen-macros` crate with `FromCsv` derive macro.
- Refactored `xiv-gen` to use manually defined structs, reducing binary size and supporting multiple languages.
- Updated `xiv-gen-db` to generate compressed bincode files for EN, JA, DE, and FR.
- Updated `ultros` backend to serve language-specific game data.
- Updated `ultros-client` to fetch game data for the user's language.
- Fixed extensive compilation errors in `ultros-app` due to struct changes.
- Improved `SubrowKey` parsing and added missing fields for various game entities.


* Support multiple language game data

- Created `xiv-gen-macros` crate with `FromCsv` derive macro.
- Refactored `xiv-gen` to use manually defined structs, reducing binary size and supporting multiple languages.
- Updated `xiv-gen-db` to generate compressed bincode files for EN, JA, DE, and FR.
- Updated `ultros` backend to serve language-specific game data.
- Updated `ultros-client` to fetch game data for the user's language.
- Fixed extensive compilation errors and clippy warnings in `ultros-app` and `ultros`.
- Improved `SubrowKey` parsing and added missing fields for various game entities.
- Cleaned up unnecessary casts and other lint warnings.


* Fix axum route syntax for game data endpoint

Updated `/static/data/{version}/{lang}` route to use `{}` syntax for path parameters, as required by Axum 0.8. This resolves a panic on server startup.


---------

+++ ultros-frontend/ultros-app/src/routes/currency_exchange.rs
@@ -881,7 +881,7 @@ pub fn CurrencySelection() -> impl IntoView {
             let Some(item) = data.items.get(f) else {
                 return false;
             };
-            allowed_item_ui_categories.contains(&ItemUiCategoryId(item.item_ui_category as i32))
+            allowed_item_ui_categories.contains(&ItemUiCategoryId(item.item_ui_category))
         })
         .unique_by(|i| i.0)
         .collect::<Vec<_>>();
